requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.6
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
kombu==4.6.8
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-http-client==3.2.7
pytz==2020.1
redis==3.5.0
sendgrid==6.3.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
